Reorganize UnusedLocalVariable and UselessAssignment
* Extracted useless setter call tracking part of UselessAssignment to
  UselessSetterCall.
* Renamed UnusedLocalVariable to UselessAssignment.
